clean up data service test structure
